+++ astropy/wcs/src/astropy_wcs.c
@@ -975,7 +975,6 @@ struct module_state {
       _setup_unit_list_proxy_type(m)||
       _setup_wcsprm_type(m)         ||
       _setup_tabprm_type(m)         ||
-      _setup_units_type(m)          ||
       /* _setup_wtbarr_type(m)         || */
       _setup_distortion_type(m)     ||
       _setup_sip_type(m)            ||