+++ astropy/wcs/tests/test_wcs.py
@@ -125,104 +125,6 @@ def test_spectra():
         # than it would have otherwise.
 
 
-def test_units():
-    with catch_warnings(AstropyDeprecationWarning) as w:
-        u = wcs.UnitConverter("log(MHz)", "ln(Hz)")
-    assert len(w) == 1
-
-    print(u.convert([1, 2, 3, 4]))
-
-basic_units = "m s g rad sr K A mol cd".split()
-derived_units = "Hz J W V N Pa C Ohm ohm S F Wb T H lm lx".split()
-add_all_units = "eV Jy R G barn".split()
-add_sup_units = "a yr pc bit byte Byte".split()
-add_sub_units = "mag".split()
-general_units = (
-    "deg arcmin arcsec mas d h min erg Ry u D DEGREE DEGREES".split())
-astro_units = "Angstrom angstrom AU lyr beam solRad solMass solLum Sun".split()
-device_units = "adu bin chan count ct photon ph pixel pix voxel".split()
-
-sub_prefixes = "y z a f p n u m c d".split()
-sup_prefixes = "da h k M G T P E Z Y".split()
-
-
-def test_all_units():
-    def test_self(x):
-        # x appears in the test name. If we would have had an ambiguous
-        # test name, we had -xxx added to the unit name.  Remove it if
-        # necessary.
-        if '-' in x:
-            x = x.split('-')[0]
-
-        # here is the test:
-        try:
-            u = wcs.UnitConverter(x, x)
-        except ValueError:
-            e = sys.exc_info()[1]
-            if str(e).startswith("ERROR 12 in wcsutrne") and \
-                    x in ("S", "H", "D"):
-                return
-            else:
-                raise
-        assert u.scale == 1.0
-        assert u.offset == 0.0
-        assert u.power == 1.0
-
-    # list of all the units to test
-    all = sorted(basic_units + derived_units + add_all_units + add_sup_units
-                 + add_sub_units + general_units + astro_units + device_units)
-
-    # Pandokia has non-case-sensitve test names; since the unit name is
-    # showing up in the test name, we want to disambiguate any name collisions.
-    # Here is a list of all the lower-cased unit name names.
-    all_lower = [x.lower() for x in all]
-
-    # here are serial numbers to use to disambiguate
-    unique_tags = {}
-
-    for unit in all:
-        # disambiguate the test name, if necessary
-        l_unit = unit.lower()
-        if unit != l_unit and l_unit in all_lower:
-            n = unique_tags.get(l_unit, 1)
-            unique_tags[n] = n + 1
-
-            # the test will tear off the part after the '-'
-            unit = '%s-%d' % (unit, n)
-
-        # perform the test
-        yield test_self, unit
-
-
-def test_unit_prefixes():
-    def test_self(x, p):
-        unit = p + x
-        try:
-            u = wcs.UnitConverter(unit, unit)
-        except ValueError:
-            e = sys.exc_info()[1]
-            if str(e) == "Potentially unsafe translation" and \
-                    x in ("S", "H", "D"):
-                return
-            else:
-                raise
-        assert u.scale == 1.0
-        assert u.offset == 0.0
-        assert u.power == 1.0
-
-    for unit in (basic_units + derived_units + add_all_units):
-        for prefix in (sub_prefixes + sup_prefixes):
-            yield test_self, unit, prefix
-
-    for unit in add_sup_units:
-        for prefix in sup_prefixes:
-            yield test_self, unit, prefix
-
-    for unit in add_sub_units:
-        for prefix in sub_prefixes:
-            yield test_self, unit, prefix
-
-
 def test_fixes():
     """
     From github issue #36
@@ -300,16 +202,6 @@ def test_load_fits_path():
     w = wcs.WCS(fits)
 
 
-def test_backward_compatible():
-    fits = get_pkg_data_filename('data/sip.fits')
-    w = wcs.WCS(fits)
-
-    with NumpyRNGContext(123456789):
-        data = np.random.rand(100, 2)
-        assert np.all(w.wcs_pix2world(data, 0) == w.wcs_pix2sky(data, 0))
-        assert np.all(w.wcs_world2pix(data, 0) == w.wcs_sky2pix(data, 0))
-
-
 def test_dict_init():
     """
     Test that WCS can be initialized with a dict-like object
@@ -349,7 +241,7 @@ def test_extra_kwarg():
     w = wcs.WCS()
     with NumpyRNGContext(123456789):
         data = np.random.rand(100, 2)
-        w.wcs_pix2sky(data, origin=1)
+        w.wcs_pix2world(data, origin=1)
 
 
 def test_3d_shapes():
@@ -359,9 +251,9 @@ def test_3d_shapes():
     w = wcs.WCS(naxis=3)
     with NumpyRNGContext(123456789):
         data = np.random.rand(100, 3)
-        result = w.wcs_pix2sky(data, 1)
+        result = w.wcs_pix2world(data, 1)
         assert result.shape == (100, 3)
-        result = w.wcs_pix2sky(
+        result = w.wcs_pix2world(
             data[..., 0], data[..., 1], data[..., 2], 1)
         assert len(result) == 3